feat: move TipoDatoStruct
