+++ src/main/java/app/MiId.java
@@ -1,29 +1,5 @@
 package app;
 
-enum TipoDato {
-  VOID,
-  INT,
-  DOUBLE,
-  FLOAT,
-  CHAR,
-  UNDEFINED;
-
-  public static TipoDato fromString(String str) {
-    if (str == null)
-      return UNDEFINED;
-    str = str.toLowerCase();
-    if (str.equals("void"))
-      return VOID;
-    if (str.equals("int"))
-      return INT;
-    if (str.equals("double"))
-      return DOUBLE;
-    if (str.equals("char"))
-      return CHAR;
-    return UNDEFINED;
-  }
-}
-
 public class MiId {
   private String token;
   private Boolean initialized;
